# docs/api/quant-radar-v1.openapi.yaml
openapi: 3.1.0
info:
  title: Surge Screener API
  version: 1.23.0-draft
  summary: Loopback-only public reads plus protected private operator state.
  description: |
    Contract for the current Phase 1 fixed-artifact reads plus bounded Phase 2
    reads for IV history, the institutional fund catalog, the manually maintained
    AI updates feed, the manually maintained schedule registry, a strict
    Options Flow and candidate feed projections, the strict Crypto Universe,
    Social Intelligence, Theme Flow, Money Flow, Sector Rotation, Daily Summary,
    Playbook Validation, Continuation Validation, COT report catalog/detail, and
    bounded Market Thesis validation/regime-history, Oversold Reversal
    validation, Reversal Radar validation, Knowledge Graph, Watchlist theme
    taxonomy, sector-to-theme drill, and Influencer roster public projections.
    Expected
    artifact failures are represented as HTTP 200 unavailable envelopes so a
    missing or half-written scheduled output cannot crash the frontend. Invalid
    request parameters use sanitized HTTP 422 problem details. Unexpected server
    defects remain HTTP 500.

    Public artifact routes intentionally have no authentication scheme because
    the server is restricted to 127.0.0.1 and exposes bounded non-sensitive
    projections. The private Industry Roles review-board routes additionally
    require a high-entropy Streamlit-to-FastAPI bearer credential. Mutations
...
servers:
  - url: http://127.0.0.1:8000
...
security: []
...
paths:
...
  /api/v1/market-context/sector-rotation/latest:
    get:
      tags: [Market Context]
      summary: Read the latest strict Sector Rotation board
      description: |
        Resolves only the newest exact dated server archive and returns the
        closed quantitative presentation board. Macro context, leader lists,
        AI read, producer status, and live providers remain server-side.
      operationId: getLatestSectorRotation
      responses:
        "200":
          $ref: "#/components/responses/SectorRotationResponse"
        "500":
          $ref: "#/components/responses/InternalProblem"
...
components:
  securitySchemes:
    InternalServiceBearer:
      type: http
      scheme: bearer
      description: |
        High-entropy Streamlit-to-FastAPI service credential for protected
        loopback routes; not a human login or provider token.
  headers:
    NoStore:
      description: Prevent browsers and intermediaries from retaining artifact state.
      schema:
        type: string
        const: no-store
...
  responses:
...
    SectorRotationResponse:
      description: Strict latest Sector Rotation presentation state.
      headers:
        Cache-Control:
          $ref: "#/components/headers/NoStore"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/SectorRotationEnvelope"
...
    InternalProblem:
      description: Unexpected programming or server failure.
      headers:
        Cache-Control:
          $ref: "#/components/headers/NoStore"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: about:blank
            title: Internal Server Error
            status: 500
  schemas:
...
    AvailableArtifactBase:
      type: object
      required: [available, reason, meta]
      properties:
        available:
          type: boolean
          const: true
        reason:
          type: string
          const: ok
        meta:
          $ref: "#/components/schemas/ArtifactMeta"
...
    AvailableSectorRotation:
      unevaluatedProperties: false
      allOf:
        - $ref: "#/components/schemas/AvailableArtifactBase"
        - type: object
          required: [data]
          properties:
            data:
              $ref: "#/components/schemas/SectorRotationData"
...
    UnavailableArtifact:
      type: object
      additionalProperties: false
      required: [available, reason, data, meta]
      properties:
        available:
          type: boolean
          const: false
        reason:
          $ref: "#/components/schemas/UnavailableReason"
        data:
          type: "null"
        meta:
          $ref: "#/components/schemas/ArtifactMeta"
    UnavailableReason:
      type: string
      enum:
        - missing
        - invalid_json
        - invalid_shape
        - unreadable
...
    SectorRotationEnvelope:
      oneOf:
        - $ref: "#/components/schemas/AvailableSectorRotation"
        - $ref: "#/components/schemas/UnavailableArtifact"
...
    SectorRotationData:
      type: object
      additionalProperties: false
      description: Strict summary-only Sector Rotation board projection.
      required: [as_of, benchmark, sectors]
      properties:
        as_of:
          type: string
          format: date
          pattern: '^\d{4}-\d{2}-\d{2}(?![\s\S])'
        benchmark:
          type: string
          const: SPY
        sectors:
          type: array
          minItems: 1
          maxItems: 100
          items:
            $ref: "#/components/schemas/SectorRotationSector"
    SectorRotationSector:
      type: object
      additionalProperties: false
      description: One closed quantitative row from the persisted Sector Rotation board.
      required:
        - etf
        - name_zh
        - group
        - theme
        - quadrant
        - quadrant_zh
        - rs_ratio
        - rs_momentum
        - heat_score
        - ret_5d
        - ret_20d
        - ret_60d
        - excess_20d
        - pct_vs_ma50
        - pct_vs_ma200
        - pct_from_52w_high
        - rvol
      properties:
        etf:
          $ref: "#/components/schemas/CandidateFeedTicker"
        name_zh:
          type: string
          minLength: 1
          maxLength: 100
        group:
          type: string
          enum: [主板塊, 主題]
        theme:
          anyOf:
            - type: string
              minLength: 1
              maxLength: 200
            - type: "null"
        quadrant:
          type: string
          enum: [Leading, Improving, Lagging, Weakening]
        quadrant_zh:
          type: string
          enum: [領漲, 醞釀, 落後, 轉弱]
        rs_ratio:
          type: number
          minimum: 0
          maximum: 500
        rs_momentum:
          type: number
          minimum: 0
          maximum: 500
        heat_score:
          type: number
          minimum: 0
          maximum: 100
        ret_5d:
          type: number
          minimum: -100
          maximum: 1000
        ret_20d:
          type: number
          minimum: -100
          maximum: 1000
        ret_60d:
          type: number
          minimum: -100
          maximum: 1000
        excess_20d:
          type: number
          minimum: -1000
          maximum: 1000
        pct_vs_ma50:
          type: number
          minimum: -100
          maximum: 1000
        pct_vs_ma200:
          type: number
          minimum: -100
          maximum: 1000
        pct_from_52w_high:
          type: number
          minimum: -100
          maximum: 1000
        rvol:
          type: number
          minimum: 0
          maximum: 1000
...
    CandidateFeedTicker:
      type: string
      minLength: 1
      maxLength: 15
      pattern: '^[A-Z0-9]+(?:[.-][A-Z0-9]+)*(?![\s\S])'
...
    ArtifactMeta:
      type: object
      additionalProperties: false
      required: [sourceId, asOf, generatedAt]
      properties:
        sourceId:
          type: string
          description: Stable registry identifier; never a filesystem path.
          pattern: "^[a-z0-9]+(?:[.-][a-z0-9]+)*$"
        asOf:
          type: [string, "null"]
          format: date
        generatedAt:
          type: [string, "null"]
          format: date-time
    Problem:
      type: object
      description: RFC 9457 problem details for unexpected server failures.
      additionalProperties: true
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          format: uri-reference